requirements.txt
botbuilder-integration-aiohttp>=4.15.0
pyodbc>=4.0.39
python-dotenv>=1.0.0